refactor(smrt-svelte)!: move agent-admin composites to smrt-agents
smrt-svelte depended on @happyvertical/smrt-agents solely for the three
agent-admin composites (AgentAdminPanel / AgentAdminTabs / AgentSettingsShell),
which type against smrt-agents' /ui registry contracts. Move them into
smrt-agents (which already hosts Svelte) so the edge disappears and smrt-svelte
drops below smrt-agents in the package DAG (unblocks #1589 universal form
adoption).

- Move the 3 .svelte composites + the tablist regression test into
  packages/agents/src/svelte/components/; rewire registry-type imports to the
  relative ../../ui.js and i18n to ../i18n.js.
- Move the 6 admin i18n keys (ui.agent_admin_panel.* / ui.agent_admin_tabs.* /
  ui.agent_settings_shell.*) into smrt-agents' svelte/i18n.ts catalog,
  preserving key names + English defaults; export from svelte/index.ts.
- Remove smrt-svelte's ./admin export, the admin barrel/dir, and the
  @happyvertical/smrt-agents dependency. Wire smrt-agents' vitest config for
  jsdom component tests (svelte plugin + browser conditions + smrt-vitest).

BREAKING CHANGE: @happyvertical/smrt-svelte/admin is removed. Import the
agent-admin shells from @happyvertical/smrt-agents/svelte instead.
