+++ packages/agents/package.json
@@ -79,6 +79,9 @@
     "@happyvertical/sql": "catalog:",
     "@sentry/node": "catalog:",
     "@sveltejs/package": "^2.5.7",
+    "@sveltejs/vite-plugin-svelte": "^6.2.4",
+    "@testing-library/svelte": "^5.3.1",
+    "@testing-library/user-event": "^14.6.1",
     "@types/node": "25.0.9",
     "fast-glob": "3.3.3",
     "svelte-check": "^4.3.5",
