+++ packages/agents/src/svelte/components/AgentAdminTabs.svelte
@@ -1,10 +1,7 @@
 <script lang="ts">
-import type {
-  AgentUIComponentRegistry,
-  AgentUISlots,
-} from '@happyvertical/smrt-agents/ui';
 import { useI18n } from '@happyvertical/smrt-ui/i18n';
-import { M } from '../../i18n/strings.workspace.js';
+import type { AgentUIComponentRegistry, AgentUISlots } from '../../ui.js';
+import { M } from '../i18n.js';
 import AgentAdminPanel from './AgentAdminPanel.svelte';
 
 const { t } = useI18n();
