+++ packages/agents/src/svelte/components/__tests__/AgentSettingsShell.tablist.test.ts
@@ -1,3 +1,4 @@
+// @vitest-environment jsdom
 /**
  * Regression: the agent switcher is a proper ARIA tablist (C8).
  *
