+++ packages/smrt-svelte/package.json
@@ -13,12 +13,6 @@
       "import": "./dist/index.js",
       "default": "./dist/index.js"
     },
-    "./admin": {
-      "types": "./dist/components/admin/index.d.ts",
-      "svelte": "./dist/components/admin/index.js",
-      "import": "./dist/components/admin/index.js",
-      "default": "./dist/components/admin/index.js"
-    },
     "./forms": {
       "types": "./dist/components/forms/index.d.ts",
       "svelte": "./dist/components/forms/index.js",
@@ -91,7 +85,6 @@
   },
   "dependencies": {
     "@happyvertical/logger": "catalog:",
-    "@happyvertical/smrt-agents": "workspace:*",
     "@happyvertical/smrt-languages": "workspace:*",
     "@happyvertical/smrt-types": "workspace:*",
     "@happyvertical/smrt-ui": "workspace:*",
